Bump SolidusSupport for combined first/last name helper
This change bumps the version of this gem in order to be able to use the
new name helper. This is necessary because in Solidus 3.0 the preference for
`use_combined_first_and_last_name_in_address` has been removed and the
default behaviour is to use the combined `name` attribute on address. We
can now use the support helper to intelligently check which attribute we
should use based on the Solidus gem version.

Co-Authored-By: MadelineCollier <madeline@super.gd>